RPN-1125 WIP

+++ enrichment/src/main/kotlin/com/rarible/protocol/union/enrichment/model/ShortOrder.kt
@@ -3,6 +3,7 @@ package com.rarible.protocol.union.enrichment.model
 import com.rarible.protocol.union.dto.BlockchainDto
 import com.rarible.protocol.union.dto.OrderIdDto
 import java.math.BigDecimal
+import java.math.BigInteger
 
 data class ShortOrder(
 
@@ -10,7 +11,7 @@ data class ShortOrder(
     val id: String,
 
     val platform: String,
-    val makeStock: BigDecimal,
+    val makeStock: BigInteger,
     val makePriceUsd: BigDecimal?,
     val takePriceUsd: BigDecimal?,
